perf: use f32 for SDF calculations and improve benchmark diagnostics

+++ src/constant.rs
@@ -1,13 +1,13 @@
-pub const GRID_SIZE: f64 = 128.;
-pub const FONT_SIZE: f64 = 96.;
-pub const BUFFER: f64 = 16.;
-pub const RADIUS: f64 = 24.;
-pub const CUTOFF: f64 = 0.25;
+pub const GRID_SIZE: f32 = 128.;
+pub const FONT_SIZE: f32 = 96.;
+pub const BUFFER: u32 = 16;
+pub const RADIUS: f32 = 24.;
+pub const CUTOFF: f32 = 0.25;
 pub const TEXTURE_SIZE: u32 = 2048;
 // 112 is just a magic number, may should be replaced by more reasonable algorithm
-pub const ASCENT: f64 = 112.;
+pub const ASCENT: f32 = 112.;
 pub const GAMMA_COEFFICIENT: f32 = 0.06;
 
 // viewport_size: (width, height)
-pub const VIEWPORT_WIDTH: f64 = 1.;
-pub const VIEWPORT_HEIGHT: f64 = 1.;
+pub const VIEWPORT_WIDTH: f32 = 1.;
+pub const VIEWPORT_HEIGHT: f32 = 1.;
